+++ Packages/Components/Sources/ListViews/Types/SelectableListType.swift
@@ -1,0 +1,17 @@
+// Copyright (c). Gem Wallet. All rights reserved.
+
+import Foundation
+
+public enum SelectableListType<T> {
+    case plain([T])
+    case section([ListSection<T>])
+    
+    public var items: [T] {
+        switch self {
+        case .plain(let items):
+            items
+        case .section(let sections):
+            sections.map { $0.values }.reduce([], +)
+        }
+    }
+}
