+++ Packages/Components/Sources/ListViews/Types/SelectionType.swift
@@ -1,0 +1,9 @@
+// Copyright (c). Gem Wallet. All rights reserved.
+
+import Foundation
+
+public enum SelectionType {
+    case multiSelection
+    case navigationLink
+    case checkmark
+}
